Reorganize Util classes
