+++ Tests/EventListener/MimeTypeListenerTest.php
@@ -24,7 +24,7 @@ class MimeTypeListenerTest extends \PHPUnit_Framework_TestCase
 {
     public function testOnKernelRequest()
     {
-        $formatNegotiator = $this->getMockBuilder('FOS\RestBundle\Util\FormatNegotiator')
+        $formatNegotiator = $this->getMockBuilder('FOS\RestBundle\Negotiation\FormatNegotiator')
             ->disableOriginalConstructor()->getMock();
         $formatNegotiator->expects($this->any())
             ->method('registerFormat')
