+++ Tests/Util/ViolationFormatterTest.php
@@ -13,7 +13,7 @@ namespace FOS\RestBundle\Tests\Util;
 
 use FOS\RestBundle\Controller\Annotations\RequestParam;
 use FOS\RestBundle\Controller\Annotations\QueryParam;
-use FOS\RestBundle\Util\ViolationFormatter;
+use FOS\RestBundle\Validator\ViolationFormatter;
 use Symfony\Component\Validator\ConstraintViolation;
 use Symfony\Component\Validator\ConstraintViolationList;
