+++ Util/Inflector/InflectorInterface.php
@@ -11,19 +11,13 @@
 
 namespace FOS\RestBundle\Util\Inflector;
 
+use FOS\RestBundle\Inflector\InflectorInterface as BaseInflectorInterface;
+
 /**
- * Inflector interface
+ * {@inheritdoc}
  *
- * @author Mark Kazemier <Markkaz>
+ * @deprecated since version 1.7, to be removed in 2.0. Use {@link BaseInflectorInterface} instead.
  */
-interface InflectorInterface
+interface InflectorInterface extends BaseInflectorInterface
 {
-    /**
-     * Pluralizes noun.
-     *
-     * @param string $word
-     *
-     * @return string
-     */
-    public function pluralize($word);
 }
