+++ Util/ViolationFormatter.php
@@ -11,38 +11,11 @@
 
 namespace FOS\RestBundle\Util;
 
-use FOS\RestBundle\Controller\Annotations\Param;
-use FOS\RestBundle\Controller\Annotations\QueryParam;
-use Symfony\Component\Validator\ConstraintViolationInterface;
-use Symfony\Component\Validator\ConstraintViolationListInterface;
+use FOS\RestBundle\Validator\ViolationFormatter as BaseViolationFormatter;
 
-class ViolationFormatter implements ViolationFormatterInterface
+/**
+ * @deprecated since version 1.7, to be removed in 2.0. Use {@link BaseViolationFormatter} instead.
+ */
+class ViolationFormatter extends BaseViolationFormatter implements ViolationFormatterInterface
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function format(Param $param, ConstraintViolationInterface $violation)
-    {
-        return sprintf(
-            "%s parameter %s value '%s' violated a constraint (%s)",
-            $param instanceof QueryParam ? 'Query' : 'Request',
-            $param->getKey(),
-            $violation->getInvalidValue(),
-            $violation->getMessage()
-        );
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function formatList(Param $param, ConstraintViolationListInterface $violationList)
-    {
-        $str = '';
-        foreach ($violationList as $key => $violation) {
-            if ($key > 0) $str .= "\n";
-            $str .= $this->format($param, $violation);
-        }
-
-        return $str;
-    }
 }
